docs: 주석 정리
