+++ src/main/java/com/sm/backend/common/response/BaseResponse.java
@@ -1,6 +1,5 @@
 package com.sm.backend.common.response;
 
-import com.sm.backend.common.exception.GlobalExceptionHandler;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
@@ -22,8 +21,6 @@ public class BaseResponse<T> {
     }
 
     public static BaseResponse fail(ResponseCode t) {
-        //return new BaseResponse(ResponseCode.FAIL.getCode(), ResponseCode.FAIL.getDefaultMessage(), t, LocalDateTime.now());
-        //return new BaseResponse(fail().getCode(), fail().getMsg(), t, LocalDateTime.now());
         return new BaseResponse(t.getCode(), t.getDefaultMessage(), t, LocalDateTime.now());
     }
