+++ src/main/java/com/sm/backend/module/address/domain/dto/ConversionCoordinatesSystem.java
@@ -1,4 +1,4 @@
-package com.sm.backend.module.address.domain.service.dto;
+package com.sm.backend.module.address.domain.dto;
 
 import com.google.gson.annotations.SerializedName;
 import lombok.AllArgsConstructor;
@@ -12,8 +12,7 @@ import java.util.List;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class AdministrativeDistrictByCoordinates {
-
+public class ConversionCoordinatesSystem {
     @SerializedName("meta")
     private Meta meta;
     @SerializedName("documents")
@@ -31,20 +30,6 @@ public class AdministrativeDistrictByCoordinates {
     @AllArgsConstructor
     @NoArgsConstructor
     public static class Document {
-        @SerializedName("region_type")
-        private String regionType;
-        @SerializedName("address_name")
-        private String addressName;
-        @SerializedName("region_1depth_name")
-        private String region1depthName;
-        @SerializedName("region_2depth_name")
-        private String region2depthName;
-        @SerializedName("region_3depth_name")
-        private String region3depthName;
-        @SerializedName("region_4depth_name")
-        private String region4depthName;
-        @SerializedName("code")
-        private String code;
         @SerializedName("x")
         private Double x;
         @SerializedName("y")
