+++ src/main/java/com/sm/backend/module/address/presentation/controller/AddressController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-@RequiredArgsConstructor // 이거 왜쓰는지
+@RequiredArgsConstructor
 @RestController
 public class AddressController {
 
@@ -18,17 +18,6 @@ public class AddressController {
     public PageResponse<SearchAddressDto.Response> search(@RequestParam("query") String query,
                                                           @RequestParam(value = "page", defaultValue = "20") int page,
                                                           @RequestParam(value = "size", defaultValue = "10") int size) {
-        // Front -> /search/adress 호출
-        // 파라미터는 동일
-        // 어떻게 kakaLocalService, Naver..Service, 행안부Service
-
-        // 위 상태로는 불가능하니까 provider (Kakao, Naver, 행안부)
-        // 적절한 방법일지 ?
-
-
-        // 카카오를 기본 -> naver/행안부
-
-        //return (Page<AddressDto.CreateResponse>) addressService.getAPI();
         return addressService.searchAddress(query, page, size);
     }
 
@@ -53,18 +42,6 @@ public class AddressController {
                                                                                           @RequestParam(value = "inputCoord", defaultValue = "WTM") String inputCoord,
                                                                                           @RequestParam(value = "outputCoord", defaultValue = "WGS84") String outputCoord) {
         return addressService.convertCoordinatesSystem(x, y, inputCoord, outputCoord);
-    }
-
-    @GetMapping("/search/mois/address")
-    public PageResponse<MoisRoadNameAddressDto.Response> searchAddress (@RequestParam(value = "confmKey", defaultValue = "devU01TX0FVVEgyMDlzMDcyMDE0NTAzOTExMzk0Njq=") String confmKey,
-                                                                        @RequestParam(value = "currentPage", defaultValue = "1") int currentPage,
-                                                                        @RequestParam(value = "countPerPage", defaultValue = "10") int countPerPage,
-                                                                        @RequestParam("keyword") String keyword,
-                                                                        @RequestParam(value = "resultType", defaultValue = "json") String resultType,
-                                                                        @RequestParam(value = "hstryYn", defaultValue = "N") String hstryYn,
-                                                                        @RequestParam(value = "firstSort", defaultValue = "NONE") String firstSort,
-                                                                        @RequestParam(value = "addInfoYn", defaultValue = "N") String addInfoYn) {
-        return addressService.searchMoisRoadNameAddress(confmKey, currentPage, countPerPage, keyword, resultType, hstryYn, firstSort, addInfoYn);
     }
 }
