--- src/main/java/com/sm/backend/module/address/presentation/dto/MoisRoadNameAddressDto.java
@@ -1,17 +1,0 @@
-package com.sm.backend.module.address.presentation.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-public class MoisRoadNameAddressDto {
-
-    @Getter
-    @Builder
-    @AllArgsConstructor
-    @NoArgsConstructor
-    public static class Response {
-        private String roadAddr;
-    }
-}
