+++ src/main/java/com/sm/backend/module/address/presentation/dto/SearchAddressDto.java
@@ -6,7 +6,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 public class SearchAddressDto {
-
     @Getter
     @Builder
     @AllArgsConstructor
