+++ src/main/java/com/sm/backend/module/manager/presentation/dto/ManagerDto.java
@@ -5,7 +5,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
-
 import javax.validation.constraints.NotNull;
 
 public class ManagerDto {
