+++ src/main/java/com/sm/backend/module/member/domain/entity/MemberPointHistory.java
@@ -8,7 +8,6 @@ import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
 import javax.persistence.*;
-import java.time.LocalDateTime;
 
 @SuperBuilder
 @AllArgsConstructor
