+++ src/main/java/com/sm/backend/module/reservation/domain/service/ReservationService.java
@@ -50,13 +50,6 @@ public class ReservationService {
                 .updatedBy("USER")
                 .build();
 
-
-        /**잔액이 부족할 때 예약이 불가능한 예외 처리
-         * 1. 포인트(point)가 얼마나 있는지 알아야하고
-         * 2. 예약 비용(rental price)가 얼마인지 알아야함.
-         * 3. 포인트가 예약 비용보다 적으면 예외 발생.
-         * */
-
         BigDecimal point = new BigDecimal(String.valueOf(reservation.getMember().getPoint()));
         BigDecimal rentalPrice = new BigDecimal(reservation.getReservableStadium().getRentalPrice());
 
@@ -68,36 +61,11 @@ public class ReservationService {
 
         member.updatePoint(remainingPoint);
 
-//        Member updatedMember = Member.builder()
-//                .id(member.getId())
-//                .reservations(member.getReservations())
-//                .memberPointHistories(member.getMemberPointHistories())
-//                .name(member.getName())
-//                .nickName(member.getNickName())
-//                .gender(member.getGender())
-//                .birth(member.getBirth())
-//                .level(member.getLevel())
-//                .manner(member.getManner())
-//                .coupon(member.getCoupon())
-//                .point(remainingPoint)
-//                .build();
-//
-//        memberRepository.save(updatedMember);
-
-    /**이미 내가 해당 구장을 예약을 했을 때
-         *1. memberId와 reservableStadiumId를 findByIds 로 조회.
-         * */
-        // 쿼리문 => select reservation_id from reservation where member_id = ? and reservable_stadium_id = ?
         boolean existMemberReservableStadiumId = reservationRepository.existsByMemberIdAndReservableStadiumId(reservation.getMember().getId(), reservation.getReservableStadium().getId());
         if (existMemberReservableStadiumId) {
             throw new ImpossibleReservationException("Already registered");
         }
 
-
-        /**정원이 가득 찼을 때
-         * 1. 현재 내가 예약하려는 구장에 몇 명이 예약했는지 알아야 함. Reservation 테이블에서 조회.
-         * 2. 구장의 최대 수용 인원을 알아야 함. Stadium ENTITY의 maximumPerson 정보를 가져옴.
-         * */
         var count = reservationRepository.findByReservationStadiumId(request.getReservableStadiumId());
         var reservationStadium = reservableStadiumRepository.findById(request.getReservableStadiumId())
                 .orElseThrow(() -> new NotFoundResourceException());
@@ -110,21 +78,12 @@ public class ReservationService {
         }
 
 
-        /**일정 티어를 넘어가면 예약 불가능.
-         * 1. 예약하려는 사람의 티어 정보를 가져와야함.
-         * 2. 예약 가능한 구장의 예약 가능한 티어 정보를 가져와야함.
-         * 3. 1,2를 비교해서 예약 가능한 티어 범위가 아니라면 예약 불가능.
-         * */
-
         Level memberLevel = reservation.getMember().getLevel();
         Level availableLevel = reservation.getReservableStadium().getLevel();
 
         if(memberLevel.getLevelPoint() > availableLevel.getLevelPoint()) {
             throw new ImpossibleReservationException("Reservation not allowed for this level");
         }
-
-        //get도 좋지만, 비교 문구 말고 compareTo() 같은 함수 호출 같은걸로도 비교해볼수 없을까?
-        //검색 키워드 => comparable, comparator (자바 기본인데 검색 후 알아보고 적용해보기)
 
         var reservedEntity = reservationRepository.save(reservation);
 
@@ -145,21 +104,11 @@ public class ReservationService {
                 .build();
     }
 
-
-    /** 예약 취소 메서드
-     * 1. 예약되어 있는 reservationId 가져오기.
-     * reservationRepository 에서 findById 메서드로 reservationId를 가져온다
-     *
-     * 2. 가져온 reservationId 의 정보가 RESERVATION 상태인지 확인하고, RESERVATION 상태가 아니라면 예외발생.
-     * 3. 예약 정보가 존재하고 RESERVATION 상태가 맞다면 CANCELED 상태로 변경.
-     * 4. updatedAt, updatedBy 상태 변경?
-     */
     @Transactional
     public ReservationDto.CancelReservationResponse cancelReservation(Long id) {
         Reservation reservation = reservationRepository.findById(id).orElseThrow(() -> new NotFoundResourceException("Not found Reservation"));
         log.info("{} reservation exists.", reservation.getId());
 
-        //예약 상태가 RESERVED 가 아니라면 예외.
         var reservationStatus = reservation.getReservationStatus();
         if(reservationStatus != ReservationStatus.RESERVED) {
             throw new NotFoundReservationException("Not found Reservation Information");
@@ -176,11 +125,11 @@ public class ReservationService {
     }
 
     public ReservationDto.FindResponse findById(Long id) {
-        Reservation reservation = reservationRepository.findById(id).orElseThrow(() -> new NotFoundResourceException("예약 정보가 없습니다."));
+        Reservation reservation = reservationRepository.findById(id).orElseThrow(() -> new NotFoundResourceException("Not found Reservation Information."));
         log.info("{} reservation exists.", reservation.getId());
 
         if (reservation == null) {
-            throw new NotFoundResourceException("예약 정보가 없습니다.");
+            throw new NotFoundResourceException("Not found Reservation Information.");
         }
 
         var findResponse = toFindResponse(reservation);
@@ -188,22 +137,8 @@ public class ReservationService {
     }
 
     private ReservationDto.FindResponse toFindResponse (Reservation entity) {
-        // getReservableStadium() 을 하는 순간 쿼리가 나가고, reservableStadium 엔티티에 값이 들어와있을 것임
-        // getStadium() 도 동일
-        // var reservableStadium = entity.getReservableStadium();
-        //var stadiumId = reservableStadium.getId();
-        //var stadiumName = stadium.getName();
-
-        //var member = entity.getMember();
-        //var memberName = member.getName();
-
-        //return ReservationDto.FindResponse.builder()
-         //       .id(entity.getId())
-           //     .build();
         var reservableStadium = entity.getReservableStadium();
         var stadiumId = reservableStadium.getId();
-
-        //var stadiumName = reservableStadium.getStadium().getName();
 
         var member = entity.getMember();
         var memberId = member.getId();
@@ -214,34 +149,4 @@ public class ReservationService {
                 .memberId(memberId)
                 .build();
     }
-
-
-
-    /**
-     * 아주 중요 X 100
-     *         1. OneToMany, ManyToOne 단방향, 양방향 매핑 방법에 대해서 공부 v
-     *         2. Entity FetchType 을 LAZY 로 되어있는 상태에서 쿼리 나가는거 확인 --> 쿼리 나온거 디코 채널에 올리기 v
-     *         3. Entity FetchType 을 EAGER 로 변경하고 쿼리 나가는거 확인 --> 쿼리 나온거 디코 채널에 올리기 v
-     *            -- 왜 LAZY 로 설정을 해야 하는지 v
-     *
-     *         4. N+1 문제를 구글링 해서 -> 이 문제가 뭔지 보고 -> 내가 작성한 코드에서 발생하는 쿼리가 N+1 문제인지 확인
-     *            (N+1 발생했다고 판단되면 디코에 올리기  => 호출할 때 발생한 쿼리와 그 코드 디코에 올리기.)
-     *         5. N+1 문제 해결하기.(여러 방법 중 어떤 방법을 선택했는지. 이유?)
-     *
-     *         6. Response 스펙에 맞게 결과 나오는거 확인
-     *         7. 검색 API 만들기(Paging + 키워드를 통한 검색).
-     */
-
-    /** 6/4
-     * 1. reserve 부분 완료해보기.
-     * 2. 코드 이해 안되는 부분 공부.
-     * 3. jpa 에서 cascade 부분 공부 orphanremoval <- 공부.
-     * */
-
-    /** 6/11
-     * 1. 예약 취소. //여러 조건들 생각해보고 주석 남기면서.. v (조금 더 수정해보기, updateRequest와 updateResponse 비교.)
-     * 2. @ExceptionHandler 키워드를 이용해서 예외가 발생했을 때, 예외 발생 포맷을 어떤 응답으로 내려줄 것인지(postman 예외 문구),,
-     * 3. manner(enum으로 수정)랑 tier도 고려해서 예약 가능하게 끔 수정. (매너가 너무 안좋으면 예약 불가, 티어가 너무 높으면 예약 불가.)
-     *    - 생각해보니 manner 점수로 예약 가능 여부를 판단하는 것은 필요없어보임.
-     * */
 }
