+++ src/main/java/com/sm/backend/module/stadium/domain/service/StadiumService.java
@@ -22,12 +22,8 @@ public class StadiumService {
     private final StadiumRepository stadiumRepository;
     private final StadiumMapper stadiumMapper;
 
-    // 상위 엔티티 = 1:N 관계에서 1인애들
-    // N+1 = 상위 엔티티 전체 조회 쿼리 1개 + 상위 엔티티 row 개수 만큼 자식 엔티티 조회 = 1 + row
     public Page<Stadium> findAll(Pageable pageable) {
-        List<Stadium> stadiums = stadiumRepository.findAll(); // 전체 조회하는 쿼리 1개
-
-        // stadium(상위 엔티티) 의 row 개수만큼 자식 연관관계에 있는 테이블을 조회함
+        List<Stadium> stadiums = stadiumRepository.findAll();
         for (var stadium: stadiums) {
             stadium.getReservableStadiums().forEach(it -> System.out.println("reservableStadiumId: " + it.getId()));
         }
