+++ src/main/java/com/sm/backend/module/stadium/infrastructure/repository/ReservableStadiumRepository.java
@@ -1,7 +1,6 @@
 package com.sm.backend.module.stadium.infrastructure.repository;
 
 import com.sm.backend.module.stadium.domain.entity.ReservableStadium;
-import com.sm.backend.module.stadium.domain.entity.Stadium;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
