+++ src/main/java/com/sm/backend/module/stadium/presentation/controller/StadiumAdminController.java
@@ -20,23 +20,12 @@ public class StadiumAdminController {
 
     private final StadiumService stadiumService;
 
-    /**
-     * 목록 페이징 + 검색
-     * 1. 페이징이 되어야 하고 -> 응답으로도 페이지 관련 필드들을 리턴해줘야함 (어떤 페이징 관련 필드를 응답으로 줄건지 설계)
-     * 2. 검색 (필터링) -> 이름 같은 조건으로 검색했을 때, 해당 데이터만 조회 되어야 함 <v 해결>
-     *
-     *
-     * 해결해야 할 것
-     * 1. BaseEntity, exception, BaseResponse, ResponseCode 코드 분석 후 적용.
-     */
-
     @GetMapping("/stadiums")
     public Page<Stadium> read() {
         PageRequest pageRequest = PageRequest.of(0,5);
         return stadiumService.findAll(pageRequest);
     }
 
-    // TODO page, size 말고 Pageable 로 사용해보기 + 추가로 정렬 조건까지 queryString 으로 주기
     @GetMapping("/stadiums/search")
     public Page<Stadium> search(@RequestParam("keyword") String keyword,
                                 @RequestParam(value = "page", defaultValue = "0") int page,
